Move watchdog.py into RimSort.utils
